Split firefox in different files matching the os
