+++ apps/firefox/firefox.py
@@ -18,24 +18,21 @@ and app.bundle: org.mozilla.firefox
 """
 
 ctx.matches = r"""
+os: windows
+os: linux
+os: mac
+tag: browser
 app: firefox
 """
-
-cmd_ctrl = "cmd" if app.platform == "mac" else "ctrl"
 
 
 @mod.action_class
 class Actions:
     def firefox_bookmarks_sidebar():
         """Toggles the Firefox bookmark sidebar"""
-        actions.key(f"{cmd_ctrl}-b")
 
     def firefox_history_sidebar():
         """Toggles the Firefox history sidebar"""
-        if app.platform == "mac":
-            actions.key("cmd-shift-h")
-        else:
-            actions.key("ctrl-h")
 
 
 @ctx.action_class("user")
@@ -47,19 +44,6 @@ class UserActions:
 
 @ctx.action_class("browser")
 class BrowserActions:
-    def bookmarks():
-        actions.key(f"{cmd_ctrl}-shift-o")
-
-    def focus_address():
-        if app.platform == "mac":
-            actions.key("cmd-l")
-        else:
-            # Only using "ctrl-l" might fail and clear the console if the user
-            # is focused in the devtools
-            actions.key("f6")
-            actions.sleep("100ms")
-            actions.key("ctrl-l")
-
     def focus_page():
         actions.browser.focus_address()
         actions.edit.find()
@@ -68,21 +52,3 @@ class BrowserActions:
 
     def go_home():
         actions.key("alt-home")
-
-    def open_private_window():
-        actions.key(f"{cmd_ctrl}-shift-p")
-
-    def show_downloads():
-        if app.platform == "linux":
-            actions.key("ctrl-shift-y")
-        else:
-            actions.key(f"{cmd_ctrl}-j")
-
-    def show_extensions():
-        actions.key(f"{cmd_ctrl}-shift-a")
-
-    def show_history():
-        if app.platform == "mac":
-            actions.key("cmd-y")
-        else:
-            actions.key("ctrl-shift-h")
